Splits types/wf.py into multiple files
